feat: fully adds sockets in notifications changes visuals and move receiver code to signals
