+++ fulabra_app/templates/fulabra_app/inbox.html
@@ -12,10 +12,10 @@
         <div class="d-flex justify-content-between align-items-center">
           
           <div>
-            <span class="badge bg-secondary mb-1 rounded-pill small">{{ note.get_notification_type_display }}</span>
+            <span class="badge bg-secondary mb-1 rounded-2 small">{{ note.get_notification_type_display }}</span>
             <p class="mb-0 fw-medium">
               {% if note.notification_type == 'friend_request' %}
-                <span class="text-primary fw-bold">{{ note.sender.username }}</span> sent you a friend request!
+                <a href="{% url "profile" note.sender.username %}" class="text-decoration-none">{{ note.sender.username }}</a> sent you a friend request!
               {% endif %}
             </p>
           </div>
@@ -26,7 +26,7 @@
                   hx-swap="outerHTML">
               {% csrf_token %}
               <input type="hidden" name="action" value="accept">
-              <button type="submit" class="btn btn-sm btn-success rounded-pill px-3 fw-bold">Accept</button>
+              <button type="submit" class="btn btn-sm btn-success rounded-4 px-3 fw-bold">Accept</button>
             </form>
 
             <form hx-post="{% url 'notification_action' note.id %}" 
@@ -34,7 +34,7 @@
                   hx-swap="outerHTML">
               {% csrf_token %}
               <input type="hidden" name="action" value="reject">
-              <button type="submit" class="btn btn-sm btn-outline-danger rounded-pill px-3 fw-bold">Decline</button>
+              <button type="submit" class="btn btn-sm btn-outline-danger rounded-4 px-3 fw-bold">Decline</button>
             </form>
           </div>
